SECURITY in user

+++ UserService/src/main/java/com/example/UserService/Controller/UserController.java
@@ -4,11 +4,14 @@ import com.example.UserService.Modules.User;
 import com.example.UserService.Service.UserService;
 import com.example.UserService.Security.JwtUtil;
 import com.example.UserService.dto.*;
+import com.example.UserService.exception.ForbiddenException;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 
 @RestController
 @RequestMapping("/users")
@@ -54,7 +57,22 @@ public class UserController {
 
     @DeleteMapping("/delete/{id}")
     public ResponseEntity<?> deleteUser(@PathVariable Long id) {
+
+        // 1️ Get logged-in user
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        String email = auth.getName();
+
+        // 2 Fetch user
+        User loggedInUser = userService.getUserByEmail(email);
+
+        // 3️ Authorization check
+        if (!loggedInUser.getUserId().equals(id)) {
+            throw new ForbiddenException("You cannot delete another user");
+        }
+
+        // 4️ Delete
         userService.deleteUser(id);
+
         return ResponseEntity.noContent().build();
     }
 }
